http_parser: Inline hints to avoid performance oddities
After the change to the iter::Iter to use the EndState enum instead of
the State enum inlining changed for the noop error, causing the
performance to degrade drastically (0.98 s to 1.4 s for the 200 MB
benchmark). The larger Error was not affected but suddenly performed
better than expected (~1.05 s).

Adding the inline(always) hint seems to produce the same result for both
versions and their performance is now back to the expected normal of
~0.98 s and ~1.08 s for the noop error and larger Error, respectively.